requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
psycopg[binary]==3.1.20
sendgrid==6.11.0
